lcbo_ca: rewrite spider to avoid embedded Google Maps parsing
Enhances AmastyStoreLocatorSpider to allow for extraction of features
which are provided only in paginated format (with e.g. maximum of 10
features returned per page).

Enhances AmastyStoreLocatorSpider to allow post_process_item to yield
requests to additional URLs to allow extraction of additional data which
is not present in the Amasty Store Locator feature response.