[Pandora] Switch to centroid search to fix spider
Previously, the single request was causing an internal server error. My assumption is that it's a timeout, i.e. too many results. I changed only the Pandora spider and not the base Rio SEO storefinder because I don't know if lat/lon search works for all brands.

Also take this opportunity to clean up the country, name, and website tags.