(PUDO parcel lockers JP) add spider
this was built by a company called GOGA so their other storefinders should be similar. https://www.goga.co.jp/works check the bottom of the page for all the brands. But there is no central URL so making a storefinder doesn't make sense, and it is such a simple format that DictParser gets most of it.