Implement Camoufox spiders with Cloudflare CAPTCHA bypass
Implement a CamoufoxSpider which can be inherited by a spider to cause
all requests to be made using Camoufox, a variant of Firefox designed
specifically for web scraping whilst avoiding bot blocking. Also
implement a Cloudflare Turnstile CAPTCHA bypass mechanism, one of the
most common mechanisms in use.

Also implement a PlaywrightSpider class to generally improve the ease of
developing Playwright-using spider. The legacy method of using spider
attribute "is_playwright_spider" is kept in place for now, pending
conversion of spiders with this attribute to inherit from
PlaywrightSpider or CamouSpider.