bmo: split spider, use w2gi domains, change w2gi to accept one key only
For the W2GI storefinder, this was modified earlier to accept more than
one API key per spider, but documentation for the spider wasn't updated
to reflect the allowance for more than one API key to be specified.

It is better to split the only spider this applied to (BMO) so that US
and CA spiders (each with their own API key) are split, and each having
their own single API key and brand name specified. This way, if the US
division of BMO switches from W2GI to some other storefinder, but the CA
division doesn't, there is only one broken spider not two.